Convert to barebones Jekyll blog site
Add the core files for a Jekyll site with `jekyll new --force .`
This puts a basic template of files in place.  The current
set moves to using the Jekyll minima layout for github pages
without any blog posts, additional pages or gem dependencies.

Can be run with `jekyll serve`

Removed the index.html to expose the rendered minima theme.